2-word-segmentation/requirements.txt
peewee==2.8.5
python-dateutil==2.5.3
snownlp==0.12.3
